Move MDX dependency from blocks and cleanup theme
