+++ addons/docs/src/DocsRenderer.tsx
@@ -1,8 +1,17 @@
-import React, { ComponentType } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import { AnyFramework, Parameters } from '@storybook/csf';
 import { DocsContextProps, DocsRenderFunction } from '@storybook/preview-web';
-import { Docs, DocsPage } from '@storybook/blocks';
+import { components as htmlComponents } from '@storybook/components';
+import { Docs, CodeOrSourceMdx, AnchorMdx, HeadersMdx } from '@storybook/blocks';
+import { MDXProvider } from '@mdx-js/react';
+
+const defaultComponents = {
+  ...htmlComponents,
+  code: CodeOrSourceMdx,
+  a: AnchorMdx,
+  ...HeadersMdx,
+};
 
 export class DocsRenderer<TFramework extends AnyFramework> {
   public render: DocsRenderFunction<TFramework>;
@@ -20,7 +29,9 @@ export class DocsRenderer<TFramework extends AnyFramework> {
       //   TODO: do we still need this? It was needed for angular (legacy) inline rendering:
       //   https://github.com/storybookjs/storybook/pull/16149
       ReactDOM.render(
-        <Docs key={Math.random()} context={context} parameters={parameters} />,
+        <MDXProvider components={defaultComponents}>
+          <Docs key={Math.random()} context={context} parameters={parameters} />
+        </MDXProvider>,
         element,
         callback
       );
